+++ examples/TuneExamplesAutoloader.php
@@ -34,12 +34,13 @@
  * @author    Jeff Tanner <jefft@tune.com>
  * @copyright 2014 Tune (http://www.tune.com)
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   0.9.8
- * @link      https://developers.mobileapptracking.com Tune Developer Community @endlink
+ * @version   0.9.9
+ * @link      https://developers.mobileapptracking.com @endlink
  *
  * Autoloader for Tune MobileAppTracking Management API files.
  *
  */
+
 namespace Tune\Examples;
 
 /**